last, i hope

+++ src/bird/Duck.java
@@ -3,7 +3,7 @@ import animal.Animal;
 import data.ColorData;
 
 public class Duck extends Animal implements iFlying {
-    public Duck(String name, int age, double weight, ColorData color) {
+    public Duck(String name, double age, double weight, ColorData color) {
         super(name, age, weight, color);
     }
