+++ src/pet/Dog.java
@@ -3,7 +3,7 @@ import animal.Animal;
 import data.ColorData;
 
 public class Dog extends Animal {
-    public Dog(String name, int age, double weight, ColorData color) {
+    public Dog(String name, double age, double weight, ColorData color) {
         super(name, age, weight, color);
     }
